+++ dart/All.hpp
@@ -36,11 +36,11 @@
 #include <dart/config.hpp>
 #include <dart/common/All.hpp>
 #include <dart/math/All.hpp>
+#include <dart/math/optimization/All.hpp>
 #include <dart/integration/All.hpp>
 #include <dart/collision/All.hpp>
 #include <dart/lcpsolver/All.hpp>
 #include <dart/constraint/All.hpp>
-#include <dart/optimizer/All.hpp>
 #include <dart/dynamics/All.hpp>
 #include <dart/simulation/All.hpp>
 // clang-format on